Refactor historyPage to be uniform with other pages
